pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "argos"
version = "0.1.0"
description = "ARGOS: Self-improving AI trading agent framework"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "gitpython>=3.1.40",
    "yfinance>=0.2.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]
data = [
    "pyarrow>=15.0.0",
    "matplotlib>=3.8.0",
]

[project.scripts]
argos-backtest = "src.agents.backtest_loop:main"
argos-eod = "src.agents.eod_cycle:main"
argos-data = "src.agents.eod_store:main"

[tool.ruff]
target-version = "py311"
line-length = 100
